fix(metadata): use struct field access instead of get_in for ParsedFileInfo
The code was using get_in() to access fields on ParsedFileInfo struct,
which caused "no function clause matching in Access.get/3" errors because
structs don't implement the Access behavior by default.

This was causing the Pluribus S01E03 files (and likely other files) to
fail during metadata enrichment with Access.get/3 exceptions.

Changed from:
  get_in(match_result, [:parsed_info, :season])
  get_in(match_result, [:parsed_info, :episodes])

To:
  match_result.parsed_info.season
  match_result.parsed_info.episodes

This fixes the import failures for files that were stuck as orphaned
in the database.